MNS-Studios/recoder/requirements.txt
pystray
pillow
keyboard
pyperclip
pyinstaller
